changed coder to use unsigned long long
git-svn-id: file:///usr/groups/dtg/svn/Cantag/trunk@98 02f9cdae-41fc-0310-9451-d5e721ae9d13

+++ CRCCoder.hh
@@ -2,6 +2,9 @@
  * $Header$
  *
  * $Log$
+ * Revision 1.2  2004/01/23 22:35:04  acr31
+ * changed coder to use unsigned long long
+ *
  * Revision 1.1  2004/01/23 18:18:11  acr31
  * added Matrix Tag and a test program and the beginning of the CRC coder
  *
@@ -22,17 +25,17 @@ class CRCCoder : Coder {
 private:
   int m_symbol_range;
   int m_symbol_count;
-  unsigned long m_encoded;
+  unsigned long long m_encoded;
   int m_counter;
 
 public:
   CRCCoder(int symbol_range,int symbol_count);
   
-  virtual void Set(unsigned long value);
+  virtual void Set(unsigned long long value);
   virtual unsigned int NextChunk();
   virtual void Reset();
   virtual bool LoadChunk(unsigned int chunk);
-  virtual unsigned long Decode();
+  virtual unsigned long long Decode();
 };
